feat: add runtime controls and enhanced animation modes
Add new entities for runtime control without reconfiguration:
- Animation speed number entity (0.1s-60s slider)
- Sync/staggered animation switch
- Scene refresh button

Enhance select entity with new features:
- "Off" option to turn off all lights
- "Random" option to pick a random scene
- Extracted palette exposed in state attributes
- Last scene change timestamp for automation triggers

Improve animation system:
- Add StaggeredAnimationController for organic, non-sync effects
- Use 0.1s instant transitions for snappy color changes
- Support float precision for all timing parameters

Update documentation with new entity patterns and attributes.

+++ custom_components/chameleon/light_controller.py
@@ -112,7 +112,7 @@ class LightController:
         ColorMode.XY,
     }
 
-    def __init__(self, hass: HomeAssistant, transition_time: int = DEFAULT_TRANSITION_TIME) -> None:
+    def __init__(self, hass: HomeAssistant, transition_time: float = DEFAULT_TRANSITION_TIME) -> None:
         """Initialize the light controller.
 
         Args:
@@ -165,7 +165,7 @@ class LightController:
         self,
         entity_id: str,
         color: RGBColor,
-        transition: int | None = None,
+        transition: float | None = None,
         brightness: int | None = None,
         skip_availability_check: bool = False,
     ) -> LightResult:
@@ -252,7 +252,7 @@ class LightController:
     async def apply_colors_to_lights(
         self,
         light_colors: dict[str, RGBColor],
-        transition: int | None = None,
+        transition: float | None = None,
         brightness: int | None = None,
     ) -> ApplyColorsResult:
         """Apply colors to multiple lights.
